requirements.txt
click
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug
gunicorn
psycopg2-binary
requests
Pillow